backend/requirements.txt
fastapi
uvicorn
playwright
google-generativeai
python-dotenv
websockets
